add assign roles to big view
